pyproject.toml
[project]
name = "sgi-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
